+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Tools/Core/InstrumentedTool.cs
@@ -5,6 +5,7 @@ using System.Text.Json;
 using ModelContextProtocol.Protocol;
 using ModelContextProtocol.Server;
 using Azure.Sdk.Tools.Cli.Telemetry;
+using Azure.Sdk.Tools.Cli.Helpers;
 using static Azure.Sdk.Tools.Cli.Telemetry.TelemetryConstants;
 
 namespace Azure.Sdk.Tools.Cli.Tools.Core;
@@ -13,6 +14,7 @@ public class InstrumentedTool : DelegatingMcpServerTool
 {
     private readonly ITelemetryService telemetryService;
     private readonly ILogger logger;
+    private readonly IMcpServerContextAccessor mcpServerContextAccessor;
     private readonly McpServerTool innerTool;
     private readonly JsonSerializerOptions serializerOptions = new()
     {
@@ -22,10 +24,12 @@ public class InstrumentedTool : DelegatingMcpServerTool
     public InstrumentedTool(
         ITelemetryService telemetryService,
         ILogger logger,
+        IMcpServerContextAccessor mcpServerContextAccessor,
         McpServerTool innerTool
     ) : base(innerTool)
     {
         this.telemetryService = telemetryService;
+        this.mcpServerContextAccessor = mcpServerContextAccessor;
         this.logger = logger;
         this.innerTool = innerTool;
     }
@@ -34,6 +38,7 @@ public class InstrumentedTool : DelegatingMcpServerTool
 
     public override async ValueTask<CallToolResult> InvokeAsync(RequestContext<CallToolRequestParams> request, CancellationToken ct = default)
     {
+        mcpServerContextAccessor.Initialize(request?.Server);
         using var activity = await telemetryService.StartActivity(ActivityName.ToolExecuted, request?.Server?.ClientInfo);
         Activity.Current = activity;
         if (request?.Params == null || string.IsNullOrEmpty(request.Params.Name))
@@ -83,14 +88,13 @@ public class InstrumentedTool : DelegatingMcpServerTool
                                         default:
                                             activity?.SetCustomProperty(kvp.Key, JsonSerializer.Serialize(kvp.Value));
                                             break;
-                                    }                                    
+                                    }
                                 }
                                 else
                                 {
                                     activity?.SetCustomProperty(kvp.Key, JsonSerializer.Serialize(kvp.Value));
                                 }
-
-                            }                  
+                            }
                         }
                     }
                 }
@@ -101,6 +105,7 @@ public class InstrumentedTool : DelegatingMcpServerTool
                 activity?.SetStatus(ActivityStatusCode.Error, ex.Message);
                 logger.LogError(ex, "Failed to deserialize contentBlock.Text for telemetry properties");
             }
+
             return result;
         }
         catch (Exception ex)